+++ deps/AppCore/src/win/MonitorWin.cpp
@@ -10,22 +10,22 @@ double MonitorWin::scale() const {
   return util_->GetMonitorDPI(monitor_);
 }
 
-int MonitorWin::width() const {
+uint32_t MonitorWin::width() const {
   MONITORINFO info;
   info.cbSize = sizeof(info);
   if (!GetMonitorInfo(monitor_, &info))
     MessageBoxW(NULL, (LPCWSTR)L"GetMonitorInfo failed", (LPCWSTR)L"Notification", MB_OK);
 
-  return abs(info.rcMonitor.left - info.rcMonitor.right);
+  return (uint32_t)abs(info.rcMonitor.left - info.rcMonitor.right);
 }
 
-int MonitorWin::height() const {
+uint32_t MonitorWin::height() const {
   MONITORINFO info;
   info.cbSize = sizeof(info);
   if (!GetMonitorInfo(monitor_, &info))
     MessageBoxW(NULL, (LPCWSTR)L"GetMonitorInfo failed", (LPCWSTR)L"Notification", MB_OK);
 
-  return abs(info.rcMonitor.top - info.rcMonitor.bottom);
+  return (uint32_t)abs(info.rcMonitor.top - info.rcMonitor.bottom);
 }
 
 }  // namespace ultralight
